added JRE-detection
With JDK7/8 it was easy to find some JRE (the JRE inside the JDK was used by javapackager/javafx-maven-plugin), but starting with JDK9+ this changed, but creating some application bundle might require some JRE, therefor we need to have this provided. Using jmods-detection within the JDK-folder, we try to choose the JRE for both worlds.